requirements.txt
streamlit
PyPDF2
langchain
sentence-transformers
faiss-cpu
google-generativeai
